Move custom code from Mod to index
