+++ src/custom/components/swap/TradeSummary/TradeSummaryMod.tsx
@@ -1,12 +1,12 @@
-import { Trans } from '@lingui/macro'
-import React, { useContext, useMemo } from 'react'
+// import { Trans } from '@lingui/macro'
+import React, { useContext /*useMemo*/ } from 'react'
 import { ThemeContext } from 'styled-components'
 import { CurrencyAmount, Currency, TradeType } from '@uniswap/sdk-core'
 
-import { Field } from 'state/swap/actions'
+// import { Field } from 'state/swap/actions'
 import { TYPE } from 'theme'
-import { computeSlippageAdjustedAmounts } from 'utils/prices'
-import { getMinimumReceivedTooltip } from 'utils/tooltips'
+// import { computeSlippageAdjustedAmounts } from 'utils/prices'
+// import { getMinimumReceivedTooltip } from 'utils/tooltips'
 
 import { AutoColumn } from 'components/Column'
 import { RowBetween, RowFixed } from 'components/Row'
@@ -14,9 +14,7 @@ import TradeGp from 'state/swap/TradeGp'
 import { MouseoverTooltipContent } from 'components/Tooltip'
 import { StyledInfo } from 'pages/Swap/SwapMod'
 import { formatSmart } from 'utils/format'
-import { TradeSummaryProps } from '.'
-import { INPUT_OUTPUT_EXPLANATION } from 'constants/index'
-import { Percent } from '@uniswap/sdk-core'
+import { RowReceivedAfterSlippage, RowSlippage, TradeSummaryProps } from '.'
 
 // computes price breakdown for the trade
 export function computeTradePriceBreakdown(trade?: TradeGp | null): {
@@ -38,44 +36,6 @@ export function computeTradePriceBreakdown(trade?: TradeGp | null): {
 export const FEE_TOOLTIP_MSG =
   'On CowSwap you sign your order (hence no gas costs!). The fees are covering your gas costs already.'
 
-export interface RowSlippageProps {
-  allowedSlippage: Percent
-  fontWeight?: number
-  fontSize?: number
-  rowHeight?: number
-}
-export function RowSlippage({ allowedSlippage, fontSize = 14, fontWeight = 500, rowHeight }: RowSlippageProps) {
-  const theme = useContext(ThemeContext)
-
-  return (
-    <RowBetween height={rowHeight}>
-      <RowFixed>
-        <TYPE.black fontSize={fontSize} fontWeight={fontWeight} color={theme.text2}>
-          <Trans>Slippage tolerance</Trans>
-        </TYPE.black>
-        <MouseoverTooltipContent
-          bgColor={theme.bg3}
-          color={theme.text1}
-          content={
-            <Trans>
-              <p>Your slippage is MEV protected: all orders are submitted with tight spread (0.1%) on-chain.</p>
-              <p>
-                The slippage you pick here enables a resubmission of your order in case of unfavourable price movements.
-              </p>
-              <p>{INPUT_OUTPUT_EXPLANATION}</p>
-            </Trans>
-          }
-        >
-          <StyledInfo />
-        </MouseoverTooltipContent>
-      </RowFixed>
-      <TYPE.black textAlign="right" fontSize={fontSize} color={theme.text1}>
-        {allowedSlippage.toFixed(2)}%
-      </TYPE.black>
-    </RowBetween>
-  )
-}
-
 export default function TradeSummary({
   trade,
   allowedSlippage,
@@ -85,13 +45,7 @@ export default function TradeSummary({
   const theme = useContext(ThemeContext)
   // const { priceImpactWithoutFee, realizedLPFee } = computeTradePriceBreakdown(trade)
   const { /* priceImpactWithoutFee, */ realizedFee } = React.useMemo(() => computeTradePriceBreakdown(trade), [trade])
-  const isExactIn = trade.tradeType === TradeType.EXACT_INPUT
-  const slippageAdjustedAmounts = computeSlippageAdjustedAmounts(trade, allowedSlippage)
-
-  const [slippageOut, slippageIn] = useMemo(
-    () => [slippageAdjustedAmounts[Field.OUTPUT], slippageAdjustedAmounts[Field.INPUT]],
-    [slippageAdjustedAmounts]
-  )
+  // const isExactIn = trade.tradeType === TradeType.EXACT_INPUT
 
   return (
     <AutoColumn gap="2px">
@@ -152,37 +106,18 @@ export default function TradeSummary({
         </RowBetween> 
         */}
 
+      {/* Slippage */}
       <RowSlippage allowedSlippage={allowedSlippage} fontSize={12} fontWeight={400} rowHeight={24} />
 
-      <RowBetween height={24}>
-        <RowFixed>
-          <TYPE.black fontSize={12} fontWeight={400} color={theme.text2}>
-            {trade.tradeType === TradeType.EXACT_INPUT ? (
-              <Trans>Minimum received (incl. fee)</Trans>
-            ) : (
-              <Trans>Maximum sent (incl. fee)</Trans>
-            )}
-          </TYPE.black>
-          {showHelpers && (
-            <MouseoverTooltipContent
-              content={getMinimumReceivedTooltip(allowedSlippage, isExactIn)}
-              bgColor={theme.bg1}
-              color={theme.text1}
-            >
-              <StyledInfo />
-            </MouseoverTooltipContent>
-          )}
-        </RowFixed>
-
-        <TYPE.black textAlign="right" fontSize={12} color={theme.text1}>
-          {/* {trade.tradeType === TradeType.EXACT_INPUT
-                ? `${trade.minimumAmountOut(allowedSlippage).toSignificant(6)} ${trade.outputAmount.currency.symbol}`
-                : `${trade.maximumAmountIn(allowedSlippage).toSignificant(6)} ${trade.inputAmount.currency.symbol}`} */}
-          {isExactIn
-            ? `${formatSmart(slippageOut) || '-'} ${trade.outputAmount.currency.symbol}`
-            : `${formatSmart(slippageIn) || '-'} ${trade.inputAmount.currency.symbol}`}
-        </TYPE.black>
-      </RowBetween>
+      {/* Min/Max received */}
+      <RowReceivedAfterSlippage
+        trade={trade}
+        showHelpers={showHelpers}
+        allowedSlippage={allowedSlippage}
+        fontSize={12}
+        fontWeight={400}
+        rowHeight={24}
+      />
 
       {/* <RowBetween>
         <RowFixed>
